Replace previous pattern matcher
For now we keep the previous one around as PatternMatcherOld.scala
